requirements.txt
fastapi
uvicorn
openai
pillow
python-dotenv
python-multipart
gunicorn